Move Guide and Plane classes into numpy_-prefixed files.
